refacto(organization): out of legacy
