+++ src/common/schema.ts
@@ -10,7 +10,7 @@ export const Entity_Schema = z.object({
 });
 
 export const Pagination_Schema = z.object({
-  page: z.string().pipe(z.coerce.number()).default("0"),
+  page: z.string().pipe(z.coerce.number()).default("1"),
   page_size: z.string().pipe(z.coerce.number()).default("10"),
 });
 export type Pagination = z.infer<typeof Pagination_Schema>;
