+++ src/organizations/:id/route.tsx
@@ -1,19 +1,39 @@
 //
 
+import type { UserInfo_Context } from ":auth/vip_list.guard";
+import type { Csp_Context } from ":common/csp_headers";
 import type { Htmx_Header } from ":common/htmx";
 import { Entity_Schema } from ":common/schema";
 import { mark_domain_as_verified } from ":legacy/services/mcp_admin_api";
 import { ORGANISATION_EVENTS } from ":organizations/services/event";
+import { Main_Layout, userinfo_to_username } from ":ui/layout/main";
 import { zValidator } from "@hono/zod-validator";
 import { Hono } from "hono";
+import { jsxRenderer } from "hono/jsx-renderer";
 import { z } from "zod";
-import organization_members_router from "./members/route";
+import Organization_Page from "./page";
+
+//
+
+const page_router = new Hono<UserInfo_Context & Csp_Context>()
+  .use("/", jsxRenderer(Main_Layout, { docType: true }))
+  .get(
+    "/",
+    zValidator("param", Entity_Schema),
+    async function GET({ req, render, notFound, var: { nonce, userinfo } }) {
+      const { id } = req.valid("param");
+      const username = userinfo_to_username(userinfo);
+      return render(<Organization_Page id={id} />, { nonce, username });
+    },
+  );
 
 //
 
 export default new Hono()
+  .route("", page_router)
   //
-  .route("members", organization_members_router)
+  // .route("members", organization_members_router)
+  // .route("domains", organization_domains_router)
   //
   .patch(
     "verify/:domain",
