+++ src/ui/layout/main.tsx
@@ -125,7 +125,7 @@ function Nav() {
         <li class="fr-nav__item">
           <a
             class="fr-nav__link"
-            href={api_ref("/legacy/organizations", {})}
+            href={app_hc.organizations.$url().pathname}
             target="_self"
           >
             Organisations
